requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
pymysql
python-dotenv
httpx
sendgrid
openai
